chore: publish ESM only
This change removes the build process of compiling the package to CJS,
making the packages ESM only. Since ESM is only supported in more modern
versions of Node.js, the minimal Node.js version requirement was
increased to 18. This is reflected in the `package.json` engines field
and GitHub action workflows.

The build process of downleveling the ESM code was also removed, as
modern versions of Node.js support all syntax used. As a result, all
build related dependencies could be removed.

There were some issues running the tests with the old version of Mocha
used. Instead of updating Mocha, I replaced it with the builtin Node.js
test runner.

Since new versions have issues consuming the old version 1 lockfile, it
was deleted and a new version 3 lockfile was generated.

This change was discussed in https://github.com/react-dropzone/attr-accept/pull/65#issuecomment-2367631383